requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.2
    # via streamlit
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
distro==1.9.0
    # via openai
exceptiongroup==1.3.1
    # via anyio
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.12.0
    # via openai
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.12.0
    # via altair
numpy==2.2.6
    # via
    #   pandas
    #   pydeck
    #   streamlit
openai==2.8.1
    # via -r requirements.in
packaging==25.0
    # via
    #   altair
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==12.0.0
    # via
    #   -r requirements.in
    #   streamlit
protobuf==6.33.1
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pypdf==6.4.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via streamlit
rpds-py==0.29.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via openai
streamlit==1.51.0
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   exceptiongroup
    #   openai
    #   pydantic
    #   pydantic-core
    #   pypdf
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
